Output an error if text is rendered without a font
Previously, the placeholder {-1} would be interpreted as a font ID. It
would then be looked up as an index, and the program would crash with
the C++ exception (on a 64-bit system):

vector::_M_range_check: __n (which is 18446744073709551615) >= this->size() (which is 1)

Avoid this by marking the font_id as optional, to avoid this happening
again, as well as checking before rendering. If it is not present, the
new FontNotSelected error code is used.