Allow reusing structure items
Say you have a paragraph containing a link, like this HTML:

  <p>
    A link like <a href="https://example.edu/">this one</a> can be
    embedded in a paragraph.
  </p>

The only way that I'm aware of to represent this within PDF is using a
tree where the paragraph has three children: a marked content region
between BDC...EMC, the /Link item, and another marked region.

I originally thought that the first and last child could be a structural
element with type /P, like in ISO 32000-1-2008 14.8.4.4.2 Example 2, but
it doesn't appear to work properly with accessibility tools. I haven't
tested this one exactly yet, but I will do shortly.

Previously, CapyPDF was not able to handle referring to MCIDs from a
structure item. With this patch, CapyPDF can no longer refer directly to
a structure item, but it can handle cases like that above
paragraph/link.

The only requirement this imposes is that the marked content is in
reading order. I don't know a nice way to remove this requirement at
this point.